Shorts/InstagramReels/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
instaloader>=4.10.0
click>=8.1.7
pytest>=7.4.0
pytest-cov>=4.1.0
sqlalchemy>=2.0.0
